refactor: 타입을 상위 컴포넌트로 이동 및 리팩토링
